Match InitObjectEventStateFromTemplate

+++ berry_fix/payload/include/global.fieldmap.h
@@ -183,8 +183,8 @@ struct ObjectEvent
     /*0x19*/ union __attribute__((packed)) {
         u8 as_byte;
         struct __attribute__((packed)) {
-            u16 x:4;
-            u16 y:4;
+            u8 x:4;
+            u8 y:4;
         } as_nybbles;
     }  range;
     /*0x1A*/ u8 fieldEffectSpriteId;
